feat(webpack-merge): change the config to merge mode
environment split into development and production

BREAKING CHANGE: environment split into development and production
