+++ src/index.js
@@ -1,12 +1,12 @@
 /*
  * @Author: your name
  * @Date: 2021-06-10 09:03:09
- * @LastEditTime: 2021-06-10 10:56:53
+ * @LastEditTime: 2021-06-10 16:45:25
  * @LastEditors: Please set LastEditors
  * @Description: In User Settings Edit
  * @FilePath: /webpack5/src/index.js
  */
-// console.log(1);
+console.log(1);
 /*
 
 > webpack5-study@0.0.1 build
@@ -18,4 +18,4 @@ webpack 5.38.1 compiled successfully in 98 ms
 
 */ 
 
-require('@bootstrap')
+// require('@bootstrap')
